backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-multipart==0.0.12
pillow==10.4.0
numpy==2.1.1
transformers==4.45.1
torch>=2.2.0
torchvision>=0.17.0
opencv-python==4.10.0.84
